# apps/orchestrator/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Outlook AI Orchestrator — API",
    "version": "0.1.0",
    "description": "AI Orchestrator backend for the Outlook AI Orchestrator.\n\n**AI-load minimisation.** `EmailAnalysis.source` tells the caller how an answer was produced:\n`llm` (model called now), `cache` (identical content already analysed), `precomputed`\n(analysed ahead of time by the mailbox-sync worker) or `heuristic` (rules only — a triaged\nnewsletter/notification, or the model was unavailable). `EmailAnalysis.triage.kind` says why.\n\n**Human-in-the-loop.** The AI never sends or deletes anything; every action goes through\n`/actions/propose` → human approval → `/actions/approve`."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Azure AD access token (`AUTH_MODE=aad`) or `ADMIN_API_TOKEN`."
      },
      "devHeaders": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-email",
        "description": "Dev identity (`AUTH_MODE=dev` only, refused in production)."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "correlationId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "ComplianceCheckRequest": {
        "type": "object",
        "properties": {
          "draft": {
            "type": "object",
            "properties": {
              "draftId": {
                "type": "string"
              },
              "from": {
                "$ref": "#/components/schemas/AnalyzeEmailRequest/properties/email/properties/from"
              },
              "to": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnalyzeEmailRequest/properties/email/properties/from"
                },
                "default": []
              },
              "cc": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnalyzeEmailRequest/properties/email/properties/from"
                },
                "default": []
              },
              "bcc": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnalyzeEmailRequest/properties/email/properties/from"
                },
                "default": []
              },
              "subject": {
                "type": "string",
                "default": ""
              },
              "body": {
                "type": "string",
                "default": ""
              },
              "attachments": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnalyzeEmailRequest/properties/email/properties/attachments/items"
                },
                "default": []
              },
              "sensitivityLabel": {
                "type": "string"
              },
              "isReplyAll": {
                "type": "boolean"
              }
            },
            "additionalProperties": false
          },
          "language": {
            "$ref": "#/components/schemas/AnalyzeEmailRequest/properties/language"
          }
        },
        "required": [
          "draft"
        ],
        "additionalProperties": false
      },
      "ComplianceCheckResponse": {
        "type": "object",
        "properties": {
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "code": {
                  "type": "string",
                  "enum": [
                    "external_recipient",
                    "confidential_attachment",
                    "missing_classification_label",
                    "sensitive_client_information",
                    "policy_violation",
                    "suspicious_recipient_domain",
                    "large_distribution",
                    "reply_all_external"
                  ]
                },
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "severity": {
                  "$ref": "#/components/schemas/EmailAnalysis/properties/risks/items/properties/severity"
                },
                "subject": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "code",
                "title",
                "description",
                "severity"
              ],
              "additionalProperties": false
            }
          },
          "recommendedActions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmailAnalysis/properties/suggestedActions/items"
            }
          },
          "verdict": {
            "type": "string",
            "enum": [
              "allow",
              "warn",
              "block"
            ]
          },
          "confidence": {
            "$ref": "#/components/schemas/EmailAnalysis/properties/classification/properties/confidence"
          },
          "auditId": {
            "type": "string"
          },
          "checkedAt": {
            "type": "string"
          }
        },
        "required": [
          "issues",
          "recommendedActions",
          "verdict",
          "confidence",
          "auditId",
          "checkedAt"
        ],
        "additionalProperties": false
      },
      "PhishingCheckRequest": {
        "type": "object",
        "properties": {
          "email": {
            "$ref": "#/components/schemas/AnalyzeEmailRequest/properties/email"
          }
        },
        "required": [
          "email"
        ],
        "additionalProperties": false
      },
      "PhishingCheckResponse": {
        "type": "object",
        "properties": {
          "score": {
            "$ref": "#/components/schemas/EmailAnalysis/properties/classification/properties/confidence"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "clean",
              "suspicious",
              "likely_phishing"
            ]
          },
          "indicators": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "weight": {
                  "type": "number"
                }
              },
              "required": [
                "code",
                "description",
                "weight"
              ],
              "additionalProperties": false
            }
          },
          "auditId": {
            "type": "string"
          }
        },
        "required": [
          "score",
          "verdict",
          "indicators",
          "auditId"
        ],
        "additionalProperties": false
      },
...
      "AuditStats": {
        "type": "object",
        "properties": {
          "period": {
            "type": "object",
            "properties": {
              "from": {
                "type": "string"
              },
              "to": {
                "type": "string"
              }
            },
            "required": [
              "from",
              "to"
            ],
            "additionalProperties": false
          },
          "kpis": {
            "type": "object",
            "properties": {
              "emailsSummarized": {
                "type": "integer"
              },
              "draftsGenerated": {
                "type": "integer"
              },
              "automationsProposed": {
                "type": "integer"
              },
              "automationsApproved": {
                "type": "integer"
              },
              "complianceAlerts": {
                "type": "integer"
              },
              "errorsAvoided": {
                "type": "integer"
              },
              "deltas": {
                "type": "object",
                "additionalProperties": {
                  "type": "number"
                },
                "default": {}
              }
            },
            "required": [
              "emailsSummarized",
              "draftsGenerated",
              "automationsProposed",
              "automationsApproved",
              "complianceAlerts",
              "errorsAvoided"
            ],
            "additionalProperties": false
          },
          "activityOverTime": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string"
                },
                "summaries": {
                  "type": "integer"
                },
                "drafts": {
                  "type": "integer"
                },
                "automations": {
                  "type": "integer"
                },
                "complianceAlerts": {
                  "type": "integer"
                }
              },
              "required": [
                "date",
                "summaries",
                "drafts",
                "automations",
                "complianceAlerts"
              ],
              "additionalProperties": false
            }
          },
          "actionsByType": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "count": {
                  "type": "integer"
                },
                "share": {
                  "type": "number"
                }
              },
              "required": [
                "type",
                "count",
                "share"
              ],
              "additionalProperties": false
            }
          },
          "complianceAlertsByCategory": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "category": {
                  "type": "string"
                },
                "count": {
                  "type": "integer"
                },
                "share": {
                  "type": "number"
                }
              },
              "required": [
                "category",
                "count",
                "share"
              ],
              "additionalProperties": false
            }
          },
          "automationsApprovalRate": {
            "type": "object",
            "properties": {
              "current": {
                "type": "number"
              },
              "previous": {
                "type": "number"
              }
            },
            "required": [
              "current",
              "previous"
            ],
            "additionalProperties": false
          },
          "topUsers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "displayName": {
                  "type": "string"
                },
                "actions": {
                  "type": "integer"
                }
              },
              "required": [
                "userId",
                "displayName",
                "actions"
              ],
              "additionalProperties": false
            }
          },
          "totalActions": {
            "type": "integer"
          }
        },
        "required": [
          "period",
          "kpis",
          "activityOverTime",
          "actionsByType",
          "complianceAlertsByCategory",
          "automationsApprovalRate",
          "topUsers",
          "totalActions"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "devHeaders": []
    }
  ],
  "paths": {
...
    "/compliance/check": {
      "post": {
        "tags": [
          "compliance"
        ],
        "operationId": "complianceCheck",
        "summary": "Pre-send compliance check",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceCheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceCheckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (`validation_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials (`unauthorized`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role (`forbidden`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (`rate_limited`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error (`internal_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/compliance/phishing": {
      "post": {
        "tags": [
          "compliance"
        ],
        "operationId": "phishingCheck",
        "summary": "Inbound phishing screening (heuristic, no model call)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhishingCheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhishingCheckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (`validation_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials (`unauthorized`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role (`forbidden`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (`rate_limited`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error (`internal_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/audit/stats": {
      "get": {
        "tags": [
          "audit"
        ],
        "operationId": "auditStats",
        "summary": "Audit KPIs",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditStats"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (`validation_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials (`unauthorized`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role (`forbidden`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (`rate_limited`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error (`internal_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/audit/export": {
      "get": {
        "tags": [
          "audit"
        ],
        "operationId": "auditExport",
        "summary": "Stream the audit trail as CSV",
        "description": "Streamed with keyset pagination, so a multi-year export never buffers in memory.",
        "responses": {
          "200": {
            "description": "CSV stream",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (`validation_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials (`unauthorized`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role (`forbidden`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (`rate_limited`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error (`internal_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}